python-ai/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
ultralytics>=8.2.0
vietocr>=0.3.6
